refactor(status.app): share content table rendering
