+++ apps/status.app/src/app/_components/content/table.tsx
@@ -1,4 +1,36 @@
+import {
+  Children,
+  isValidElement,
+  type ReactElement,
+  type ReactNode,
+} from 'react'
+
 import { renderText } from '~app/_utils/render-text'
+
+type ContentElement = ReactElement<{ children?: ReactNode }>
+
+const toElementArray = (children: ReactNode): ContentElement[] =>
+  Children.toArray(children).filter((child): child is ContentElement =>
+    isValidElement<{ children?: ReactNode }>(child)
+  )
+
+const isHtmlElement = (element: ContentElement, tagName: string) =>
+  element.type === tagName
+
+const getTableRows = (element?: ContentElement) => {
+  if (!element) {
+    return []
+  }
+
+  return toElementArray(element.props.children).filter(row =>
+    isHtmlElement(row, 'tr')
+  )
+}
+
+const getTableCells = (row: ContentElement) =>
+  toElementArray(row.props.children).filter(
+    cell => isHtmlElement(cell, 'td') || isHtmlElement(cell, 'th')
+  )
 
 export function Table(props: {
   hasShadow?: boolean
@@ -75,5 +107,51 @@ export function TableCell(props: { children: React.ReactNode }) {
     <td className="p-3 align-top">
       {renderText({ children, size: 'text-15' })}
     </td>
+  )
+}
+
+// Handles tables missing a <thead> (older Ghost tables) or otherwise not
+// shaped as exactly [thead, tbody], instead of assuming that fixed shape.
+export function renderContentTable(children: ReactNode) {
+  const elements = toElementArray(children)
+  const head = elements.find(child => isHtmlElement(child, 'thead'))
+  const bodies = elements.filter(child => isHtmlElement(child, 'tbody'))
+  const directRows = elements.filter(child => isHtmlElement(child, 'tr'))
+  const footer = elements.find(child => isHtmlElement(child, 'tfoot'))
+  const headerRows = getTableRows(head)
+  const headerCells = headerRows[0] ? getTableCells(headerRows[0]) : []
+  const bodyRows = [
+    ...headerRows.slice(1),
+    ...bodies.flatMap(getTableRows),
+    ...directRows,
+    ...getTableRows(footer),
+  ]
+
+  return (
+    <Table>
+      {headerCells.length > 0 && (
+        <TableHead>
+          {headerCells.map((header, index) => (
+            <TableHeader key={index}>{header.props.children}</TableHeader>
+          ))}
+        </TableHead>
+      )}
+      <TableContent>
+        {bodyRows.map((row, index) => (
+          <TableRow key={index}>
+            {getTableCells(row).map((cell, index) =>
+              // A <th> in a body row is a row header (common in tables with
+              // no <thead>); rendering it as a <td> would drop that
+              // semantic and its header styling.
+              isHtmlElement(cell, 'th') ? (
+                <TableHeader key={index}>{cell.props.children}</TableHeader>
+              ) : (
+                <TableCell key={index}>{cell.props.children}</TableCell>
+              )
+            )}
+          </TableRow>
+        ))}
+      </TableContent>
+    </Table>
   )
 }
